Добавил отображение скорости ветра, давления, влажности и переключение с цельсия на фаренгейты.

+++ WeatherAppMaterial/app/src/main/java/com/example/weatherappmaterial/data/Wind.java
@@ -6,16 +6,16 @@ import com.google.gson.annotations.SerializedName;
 public class Wind {
     @SerializedName("speed")
     @Expose
-    private Integer speed;
+    private float speed;
     @SerializedName("deg")
     @Expose
-    private Integer deg;
+    private int deg;
 
-    public Integer getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
-    public void setSpeed(Integer speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
 }
